Ok driver project/backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
python-dotenv
